+++ packages/mds-api-server/api-server.ts
@@ -47,8 +47,6 @@ export interface ApiServerOptions {
 }
 
 export const ApiServer = (
-  // The linter does not realize that the type variable is used.
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   api: (server: express.Express) => express.Express,
   options: Partial<ApiServerOptions> = {},
   app: express.Express = express()