+++ packages/mds-compliance-engine/engine/speed_processors.ts
@@ -56,8 +56,6 @@ export function processSpeedPolicy(
             rules_matched: [rule.rule_id],
             speed: event.telemetry.gps.speed as number
           }
-          /* eslint-reason need to remove matched vehicles */
-          /* eslint-disable-next-line no-param-reassign */
           delete devicesToCheck[device.device_id]
         }
       }