+++ packages/mds-policy-service/tests/policy-repository.spec.ts
@@ -1,4 +1,3 @@
-/* eslint-disable no-console */
 import type { UUID } from '@mds-core/mds-types'
 import {
   clone,
@@ -26,7 +25,6 @@ import {
 } from '../test_data/policies'
 import { PolicyFactory, RulesFactory } from './helpers'
 
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
 const ACTIVE_POLICY_JSON = { ...POLICY_JSON, publish_date: yesterday(), start_date: yesterday() }
 
 const SIMPLE_POLICY_JSON: PolicyDomainCreateModel = {
@@ -52,21 +50,10 @@ const SIMPLE_POLICY_JSON: PolicyDomainCreateModel = {
 }
 
 describe('Policy Repository Tests', () => {
-  beforeAll(async () => {
-    await PolicyRepository.initialize()
-  })
-
-  it('Run Migrations', async () => {
-    await PolicyRepository.runAllMigrations()
-  })
-
-  it('Revert Migrations', async () => {
-    await PolicyRepository.revertAllMigrations()
-  })
-
-  afterAll(async () => {
-    await PolicyRepository.shutdown()
-  })
+  beforeAll(PolicyRepository.initialize)
+  it('Run Migrations', PolicyRepository.runAllMigrations)
+  it('Revert Migrations', PolicyRepository.revertAllMigrations)
+  afterAll(PolicyRepository.shutdown)
 })
 
 describe('spot check unit test policy functions with SimplePolicy', () => {