+++ container-images/mds-agency/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @container-images/mds-agency
 
+## 0.2.1
+
+### Patch Changes
+
+- @mds-core/mds-agency@0.5.12
+
 ## 0.2.0
 
 ### Minor Changes