+++ packages/mds-db/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @mds-core/mds-db
 
+## 0.6.17
+
+### Patch Changes
+
+- ed8ecc80: Upgrade to TypeORM v3
+- Updated dependencies [ed8ecc80]
+  - @mds-core/mds-attachment-service@0.5.26
+  - @mds-core/mds-audit-service@0.6.3
+  - @mds-core/mds-geography-service@0.10.22
+  - @mds-core/mds-ingest-service@0.10.10
+  - @mds-core/mds-policy-service@0.11.28
+
 ## 0.6.16
 
 ### Patch Changes