+++ packages/mds-ingest-service/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @mds-core/mds-ingest-service
 
+## 0.10.10
+
+### Patch Changes
+
+- ed8ecc80: Upgrade to TypeORM v3
+- Updated dependencies [91159748]
+- Updated dependencies [ed8ecc80]
+  - @mds-core/mds-rpc-common@0.3.19
+  - @mds-core/mds-repository@0.2.0
+  - @mds-core/mds-schema-validators@0.5.10
+
 ## 0.10.9
 
 ### Patch Changes